Fast API/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["QP-MRMousavi <m.r.mousavi.n@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.8"
fastapi = "^0.110.2"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
tortoise-orm = {extras = ["asyncmy"], version = "^0.20.1"}
email-validator = "^2.1.1"
orjson = "^3.10.1"
aerich = "^0.7.2"
Authlib = "^1.3.0"
docformatter = "^1.7.5"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.2.1"
structlog = "^24.1.0"
rich = "^13.7.1"
httpx = "^0.27.0"
asyncmy = "^0.2.9"
asgi-lifespan = "^2.1.0"
pytest-asyncio = "^0.23.6"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
pytest-mock = "^3.8.0"
mypy = "^1.9.0"
black = "^24.4.0"
isort = "^5.13.2"
autoflake = "^2.3.1"

[tool.coverage.run]
omit = ["app/tests*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:"
    ]

[tool.black]
line-length = 90
target-version=[
    "py311"
    ]

[tool.isort]
quiet=true
profile = "black"

[tool.docformatter]
recursive = true
wrap-summaries = 90
wrap-descriptions = 90
blank = true
close-quotes-on-newline = true

[tool.autoflake]
recursive = true
expand-star-imports = true
remove-all-unused-imports = true
ignore-init-module-imports = true
remove-duplicate-keys = true
remove-unused-variables = true

[tool.mypy]
plugins = [
    "pydantic.mypy"
    ]
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
no_implicit_reexport = true
ignore_missing_imports = true
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[tool.aerich]
tortoise_orm = "app.core.configs.TORTOISE_CONFIG"
location = "./migrations"
src_folder = "./"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
